+++ src/Writable.ts
