+++ src/index.ts
@@ -1,4 +1,5 @@
 export { BufferType } from "./BufferType";
+export { Writable } from "./Writable";
 export { encode } from "./encode";
 export { decode } from "./decode";
 export { ExtensionCodec, ExtensionCodecType } from "./ExtensionCodec";
